Further cleanups to remaining C++ code.
