+++ src/dev_tools/testRunner/testCases/tools/ethNames.csv
@@ -7,6 +7,7 @@ on      ,both ,fast  ,names ,tools/ethNames ,invalid_param_3        ,n    ,terms
 on      ,both ,medi  ,names ,tools/ethNames ,fmt_default            ,y    ,terms = tip
 on      ,both ,medi  ,names ,tools/ethNames ,fmt_txt                ,n    ,terms = tip & fmt = txt
 on      ,both ,medi  ,names ,tools/ethNames ,fmt_csv                ,n    ,terms = tip & fmt = csv
+on      ,both ,medi  ,names ,tools/ethNames ,fmt_csv_expand         ,n    ,terms = tip & fmt = csv & expand
 on      ,both ,medi  ,names ,tools/ethNames ,fmt_api                ,y    ,terms = tip & fmt = api
 on      ,both ,medi  ,names ,tools/ethNames ,fmt_json               ,y    ,terms = tip & fmt = json
 on      ,both ,fast  ,names ,tools/ethNames ,fmt_junk               ,y    ,terms = tip & fmt = junk
@@ -67,17 +68,17 @@ on      ,both ,fast  ,names ,tools/ethNames ,simple_addr_only       ,y    ,terms
 on      ,both ,medi  ,names ,tools/ethNames ,duplicated_option_1    ,n    ,terms = 0xb6 & terms = torj & addr & addr & addr & addr
 
 # Search for TODO: Fix storing and reading names from / to binary cache
-on      ,both ,fast  ,names ,tools/ethNames ,edit_invalid           ,y    ,invalid & log_level = 10 & fmt = json & expand
-on      ,both ,fast  ,names ,tools/ethNames ,edit_add               ,y    ,create & log_level = 10 & fmt = json & expand
-local   ,both ,fast  ,names ,tools/ethNames ,show_add               ,y    ,terms = 0x0000000000000000000000000000000010101010 & fmt = json & expand
-on      ,both ,fast  ,names ,tools/ethNames ,edit_edit              ,y    ,update & log_level = 10 & fmt = json & expand
-local   ,both ,fast  ,names ,tools/ethNames ,show_edit              ,y    ,terms = 0x0000000000000000000000000000000010101010 & fmt = json & expand
-on      ,both ,fast  ,names ,tools/ethNames ,edit_delete            ,y    ,delete & terms = 0x0000000000000000000000000000000010101010 & log_level = 10 & fmt = json & expand
-local   ,both ,fast  ,names ,tools/ethNames ,show_delete            ,y    ,terms = 0x0000000000000000000000000000000010101010 & fmt = json & expand
-on      ,both ,fast  ,names ,tools/ethNames ,edit_undelete          ,y    ,undelete & terms = 0x0000000000000000000000000000000010101010 & log_level = 10 & fmt = json & expand
-local   ,both ,fast  ,names ,tools/ethNames ,show_undelete          ,y    ,terms = 0x0000000000000000000000000000000010101010 & fmt = json & expand
-on      ,both ,fast  ,names ,tools/ethNames ,edit_remove            ,y    ,remove & terms = 0x0000000000000000000000000000000010101010 & log_level = 10 & fmt = json & expand
-local   ,both ,fast  ,names ,tools/ethNames ,show_remove            ,y    ,terms = 0x0000000000000000000000000000000010101010 & fmt = json & expand
+on      ,both ,fast  ,names ,tools/ethNames ,edit_invalid           ,y    ,invalid & log_level = 10 & fmt = json
+on      ,both ,fast  ,names ,tools/ethNames ,edit_add               ,y    ,create & log_level = 10 & fmt = json
+local   ,both ,fast  ,names ,tools/ethNames ,show_add               ,y    ,terms = 0x0000000000000000000000000000000010101010 & fmt = json
+on      ,both ,fast  ,names ,tools/ethNames ,edit_edit              ,y    ,update & log_level = 10 & fmt = json
+local   ,both ,fast  ,names ,tools/ethNames ,show_edit              ,y    ,terms = 0x0000000000000000000000000000000010101010 & fmt = json
+on      ,both ,fast  ,names ,tools/ethNames ,edit_delete            ,y    ,delete & terms = 0x0000000000000000000000000000000010101010 & log_level = 10 & fmt = json
+local   ,both ,fast  ,names ,tools/ethNames ,show_delete            ,y    ,terms = 0x0000000000000000000000000000000010101010 & fmt = json
+on      ,both ,fast  ,names ,tools/ethNames ,edit_undelete          ,y    ,undelete & terms = 0x0000000000000000000000000000000010101010 & log_level = 10 & fmt = json
+local   ,both ,fast  ,names ,tools/ethNames ,show_undelete          ,y    ,terms = 0x0000000000000000000000000000000010101010 & fmt = json
+on      ,both ,fast  ,names ,tools/ethNames ,edit_remove            ,y    ,remove & terms = 0x0000000000000000000000000000000010101010 & log_level = 10 & fmt = json
+local   ,both ,fast  ,names ,tools/ethNames ,show_remove            ,y    ,terms = 0x0000000000000000000000000000000010101010 & fmt = json
 
-local   ,both ,fast  ,names ,tools/ethNames ,edit_custom_delete     ,y    ,delete & terms = 0x054993ab0f2b1acc0fdc65405ee203b4271bebe6 & log_level = 10 & to_custom & all & expand
-local   ,both ,fast  ,names ,tools/ethNames ,edit_custom_undelete   ,y    ,undelete & terms = 0x054993ab0f2b1acc0fdc65405ee203b4271bebe6 & log_level = 10 & to_custom & all & expand
+local   ,both ,fast  ,names ,tools/ethNames ,edit_custom_delete     ,y    ,delete & terms = 0x054993ab0f2b1acc0fdc65405ee203b4271bebe6 & log_level = 10 & to_custom & all
+local   ,both ,fast  ,names ,tools/ethNames ,edit_custom_undelete   ,y    ,undelete & terms = 0x054993ab0f2b1acc0fdc65405ee203b4271bebe6 & log_level = 10 & to_custom & all
